feat(deployer): consolodate zero downtime tests into a single suite for agentgateway and envoy
