--- test/kubernetes/e2e/features/agentgateway/zero_downtime_rollout/testdata/hey.yaml
@@ -1,24 +1,0 @@
----
-apiVersion: v1
-kind: Namespace
-metadata:
-  name: hey
----
-apiVersion: v1
-kind: Pod
-metadata:
-  name: hey
-  namespace: hey
-  labels:
-    app: hey
-    version: v1
-    app.kubernetes.io/name: hey
-spec:
-  containers:
-    - name: hey
-      image: ricoli/hey
-      imagePullPolicy: IfNotPresent
-      command:
-        - "tail"
-        - "-f"
-        - "/dev/null"
