--- test/kubernetes/e2e/features/agentgateway/zero_downtime_rollout/testdata/service.yaml
@@ -1,26 +1,0 @@
----
-apiVersion: v1
-kind: Pod
-metadata:
-  name: nginx
-  labels:
-    app.kubernetes.io/name: nginx
-spec:
-  containers:
-    - name: nginx
-      image: nginx:stable
-      ports:
-        - containerPort: 80
-          name: http-web-svc
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: example-svc
-spec:
-  selector:
-    app.kubernetes.io/name: nginx
-  ports:
-    - protocol: TCP
-      port: 8080
-      targetPort: http-web-svc
