+++ test/kubernetes/e2e/features/zero_downtime_rollout/suite.go
@@ -51,8 +51,8 @@ func (s *testingSuite) TestZeroDowntimeRollout() {
 	// Run this for 30s which is long enough to restart the deployment since there's no easy way
 	// to stop this command once the test is over.
 	// This executes 800 req @ 4 req/sec = 20s (3 * terminationGracePeriodSeconds (5) + buffer).
-	// kubectl exec -n hey hey -- hey -disable-keepalive -c 4 -q 10 --cpus 1 -n 800 -m GET -t 1 -host example.com http://gw.default.svc.cluster.local:8080.
-	args := []string{"exec", "-n", "hey", "hey", "--", "hey", "-disable-keepalive", "-c", "4", "-q", "10", "--cpus", "1", "-n", "800", "-m", "GET", "-t", "1", "-host", "example.com", "http://gw.default.svc.cluster.local:8080"}
+	// kubectl exec -n hey heygw -- hey -disable-keepalive -c 4 -q 10 --cpus 1 -n 800 -m GET -t 1 -host example.com http://gw.default.svc.cluster.local:8080.
+	args := []string{"exec", "-n", "hey", "heygw", "--", "hey", "-disable-keepalive", "-c", "4", "-q", "10", "--cpus", "1", "-n", "800", "-m", "GET", "-t", "1", "-host", "example.com", "http://gw.default.svc.cluster.local:8080"}
 
 	cmd := kCli.Command(s.Ctx, args...)
 
@@ -75,16 +75,58 @@ func (s *testingSuite) TestZeroDowntimeRollout() {
 		s.T().Fatal("error waiting for command to finish", err)
 	}
 
-	// Command output should contain this:
-	//   Status code distribution:
-	// 		[200]	800 responses
-	//
-	// And, not contain something like this:
-	//   Status code distribution:
-	// 		[200]	779 responses
-	// 	Error distribution:
-	//   	[17]	Get http://agentgw.default.svc.cluster.local:8080: dial tcp 10.96.177.91:8080: connection refused
-	//   	[4]	Get http://agentgw.default.svc.cluster.local:8080: net/http: request canceled while waiting for connection
+	// Verify that there were no errors.
+	s.Contains(string(cmd.Output()), "[200]	800 responses")
+	s.NotContains(string(cmd.Output()), "Error distribution")
+}
+
+func (s *testingSuite) TestZeroDowntimeRolloutAgentGateway() {
+	// Ensure the agentgateway pod is up and running.
+	s.TestInstallation.Assertions.EventuallyPodsRunning(s.Ctx,
+		agentgatewayObjectMeta.GetNamespace(), metav1.ListOptions{
+			LabelSelector: "app.kubernetes.io/name=" + agentgatewayObjectMeta.GetName(),
+		})
+
+	s.TestInstallation.Assertions.AssertEventualCurlResponse(
+		s.Ctx,
+		defaults.CurlPodExecOpt,
+		[]curl.Option{
+			curl.WithHost(kubeutils.ServiceFQDN(agentgatewayObjectMeta)),
+			curl.WithHostHeader("example.com"),
+		},
+		&testmatchers.HttpResponse{
+			StatusCode: http.StatusOK,
+		})
+
+	kCli := kubectl.NewCli()
+
+	// Send traffic to the gateway pod while we restart the deployment.
+	// Run this for 30s which is long enough to restart the deployment since there's no easy way
+	// to stop this command once the test is over.
+	// This executes 800 req @ 4 req/sec = 20s (3 * terminationGracePeriodSeconds (5) + buffer).
+	// kubectl exec -n hey heyagw -- hey -disable-keepalive -c 4 -q 10 --cpus 1 -n 800 -m GET -t 1 -host example.com http://agentgw.default.svc.cluster.local:8080.
+	args := []string{"exec", "-n", "hey", "heyagw", "--", "hey", "-disable-keepalive", "-c", "4", "-q", "10", "--cpus", "1", "-n", "800", "-m", "GET", "-t", "1", "-host", "example.com", "http://agentgw.default.svc.cluster.local:8080"}
+
+	cmd := kCli.Command(s.Ctx, args...)
+
+	if err := cmd.Start(); err != nil {
+		s.T().Fatal("error starting command", err)
+	}
+
+	// Restart the deployment, twice.
+	// There should be no downtime, since the gateway pod
+	// should have readiness probes configured.
+	err := kCli.RestartDeploymentAndWait(s.Ctx, "agentgw")
+	s.Require().NoError(err)
+
+	time.Sleep(time.Second)
+
+	err = kCli.RestartDeploymentAndWait(s.Ctx, "agentgw")
+	s.Require().NoError(err)
+
+	if err := cmd.Wait(); err != nil {
+		s.T().Fatal("error waiting for command to finish", err)
+	}
 
 	// Verify that there were no errors.
 	s.Contains(string(cmd.Output()), "[200]	800 responses")
