+++ test/kubernetes/e2e/features/zero_downtime_rollout/testdata/agentgateway.yaml
@@ -27,3 +27,22 @@ spec:
     - backendRefs:
         - name: example-svc
           port: 8080
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: heyagw
+  namespace: hey
+  labels:
+    app: heyagw
+    version: v1
+    app.kubernetes.io/name: heyagw
+spec:
+  containers:
+    - name: hey
+      image: ricoli/hey
+      imagePullPolicy: IfNotPresent
+      command:
+        - "tail"
+        - "-f"
+        - "/dev/null"
