+++ test/kubernetes/e2e/features/zero_downtime_rollout/testdata/gateway.yaml
@@ -26,3 +26,22 @@ spec:
     - backendRefs:
         - name: example-svc
           port: 8080
+---
+apiVersion: v1
+kind: Pod
+metadata:
+  name: heygw
+  namespace: hey
+  labels:
+    app: heygw
+    version: v1
+    app.kubernetes.io/name: heygw
+spec:
+  containers:
+    - name: hey
+      image: ricoli/hey
+      imagePullPolicy: IfNotPresent
+      command:
+        - "tail"
+        - "-f"
+        - "/dev/null"
