+++ test/kubernetes/e2e/features/zero_downtime_rollout/testdata/service.yaml
@@ -24,3 +24,8 @@ spec:
     - protocol: TCP
       port: 8080
       targetPort: http-web-svc
+---
+apiVersion: v1
+kind: Namespace
+metadata:
+  name: hey
